chore: clean unused enums
